+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.7.5 - TBD
+ * Removed `NTLMSoapClient\Exchange`. Folded the few lines of functionality in to `NTLMSoapClient`
+ * Added `API::getServerTimezones($timezoneIDs = array(), $fullTimezoneData = false)`
+
 ## 0.7.4 - 2016-03-02
  * Introduced the `API::deleteFolder($folderId)` method
  * Improved OAuth Documentation